.github/actions/create-pull-request/requirements.txt
PyGithub==1.55.0
